src/vla_integration/requirements.txt
# ROS 2 dependencies
rclpy>=3.0.0
ros2cli>=0.9.0

# OpenAI dependencies
openai>=1.0.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
numpy>=1.21.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0